Fix info loading + refactoring
